backend/pyproject.toml
[project]
name = "raijin-backend"
version = "0.1.0"
description = "Raijin API — invoice automation backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "python-multipart>=0.0.17",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<5",
    "slowapi>=0.1.9",
    "boto3>=1.35",
    "celery[redis]>=5.4",
    "redis>=5.2",
    "httpx>=0.28",
    "python-dateutil>=2.9",
    "email-validator>=2.2",
    "structlog>=24.4",
    "openpyxl>=3.1.5",
    "cryptography>=43.0",
    "sentry-sdk[fastapi]>=2.19",
    "msal>=1.31",
    "google-auth>=2.36",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.150",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-python-jose",
    "types-passlib",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "RUF001", "RUF002", "RUF003"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
